Moved the group-by-owner table and helper functions into a separate component.
